+++ src/onegov.agency/onegov/agency/views/settings.py
@@ -2,16 +2,12 @@ from onegov.agency import _
 from onegov.agency.app import AgencyApp
 from onegov.core.security import Secret
 from onegov.form import Form
-from onegov.form import merge_forms
-from onegov.org.forms import SettingsForm
 from onegov.org.models import Organisation
-from onegov.org.views.settings import handle_settings
+from onegov.org.views.settings import handle_generic_settings
 from wtforms import RadioField
 
 
-def settings_form(model, request):
-
-    class CustomFieldsForm(Form):
+class PDFLayoutForm(Form):
 
         pdf_layout = RadioField(
             label=_("PDF Layout"),
@@ -32,16 +28,15 @@ def settings_form(model, request):
             super().populate_obj(obj, *args, **kwargs)
             obj.meta['pdf_layout'] = self.pdf_layout.data
 
-    return merge_forms(SettingsForm, CustomFieldsForm)
-
 
 @AgencyApp.form(
     model=Organisation,
-    name='settings',
+    name='pdf-layout-settings',
     template='form.pt',
     permission=Secret,
-    form=settings_form
+    form=PDFLayoutForm,
+    setting=_("PDF Layout"),
+    icon='fa-file-pdf-o'
 )
-def custom_handle_settings(self, request, form):
-    form.delete_field('default_map_view')
-    return handle_settings(self, request, form)
+def handle_pdf_layout_settings(self, request, form):
+    return handle_generic_settings(self, request, form, _("PDF Layout"))