+++ src/onegov.agency/HISTORY.rst
@@ -1,5 +1,9 @@
 Changelog
 ---------
+
+- Reorganises settings into separate views.
+  [href]
+
 1.5.1 (2019-01-07)
 ~~~~~~~~~~~~~~~~~~~~